Add rad-auth-keys CLI tool
The `rad-auth-keys` CLI tool manages a proposed
`<git-repo>/.rad/keys` file for maintaining a keyring for namespace
access control.

This commit is part of a proof of concept for improving access control
to repositories through the use of validating git signed push
certificates evaluated in the `pre-receive` hook for the
`radicle-git-server`.

The intent of the `.rad/keys` directory is to live within the root of a
git repository, and to be updated with maintainers' GPG key fingerprints
to authorize a signed push request.

This means access control is managed by the maintainers accepting
changes to the `.rad/keys` directory.

Additionally, the structure of the `.rad/keys/` directory is made to be
consistent with other key management schemes, namely `patatt`.

Signed-off-by: Alexis Sellier <self@cloudhead.io>