Print the version only, without the name
Before this change, invoking the CLI as `rad --version` would print something
like `rad 0.6.1`. Notice that the output is not valid semver because it contains
the `rad ` substring preceding the actual version number.

Because of this, clients of the CLI that want to interpolate the CLI version
in a composed string e.g.:

```js
const log = `Using rad CLI v${getRadCliVersion()} from ${getRadCliPath()}`
```

will end up with the following as a result:

`Using rad CLI vrad 0.6.1 from /usr/bin/rad`

Note how we wanted to have `v0.6.1` and instead have `vrad 0.6.1`.

This commit fixes the above issue for the specific command `rad --version`
while taking care that there are no side-effects to the output of other other
existing commands.

Signed-off-by: Konstantinos Maninakis <maninak@protonmail.com>